major changes to IO modes to allow for 64 bit IDs so we can scale to almost 2.1 billion elements (have to check the exact number).

+++ src/Blocks.jl
@@ -21,6 +21,7 @@ print(io, "Block:\n",
 
 # more verbose types
 #
+# TODO turn this into actual types with constructors to declutter stuff
 BlockIDs = Vector{BlockID}
 BlockIDsPtr = Ref{BlockID}
 Blocks = Vector{Block}
@@ -61,14 +62,21 @@ function read_block_connectivity(exo_id::ExoID, block_id::BlockID)
     element_type, num_elem, num_nodes, num_edges, num_faces, num_attributes =
     read_element_block_parameters(exo_id::ExoID, block_id::BlockID)
 
-    conn = Vector{Int32}(undef, num_nodes * num_elem)
-    conn_face = Vector{Int32}(undef, num_nodes * num_elem)  # Not using these currently
-    conn_edge = Vector{Int32}(undef, num_nodes * num_elem)  # Not using these currently
+    # old 32 bit IDs behavior
+    # conn = Vector{Int32}(undef, num_nodes * num_elem)
+    # conn_face = Vector{Int32}(undef, num_nodes * num_elem)  # Not using these currently
+    # conn_edge = Vector{Int32}(undef, num_nodes * num_elem)  # Not using these currently
+    conn = Vector{Int64}(undef, num_nodes * num_elem)
+    conn_face = Vector{Int64}(undef, num_nodes * num_elem)  # Not using these currently
+    conn_edge = Vector{Int64}(undef, num_nodes * num_elem)  # Not using these currently
 
     # TODO: look into why the connectivity arrays need to be 32 bit.
     #
+    # error = ccall((:ex_get_conn, libexodus), ExodusError,
+    #               (ExoID, ExodusConstant, BlockID, Ref{Int32}, Ref{Int32}, Ref{Int32}),
+    #               exo_id, EX_ELEM_BLOCK, block_id, conn, conn_face, conn_edge)
     error = ccall((:ex_get_conn, libexodus), ExodusError,
-                  (ExoID, ExodusConstant, BlockID, Ref{Int32}, Ref{Int32}, Ref{Int32}),
+                  (ExoID, ExodusConstant, BlockID, Ref{Int64}, Ref{Int64}, Ref{Int64}),
                   exo_id, EX_ELEM_BLOCK, block_id, conn, conn_face, conn_edge)
     exodus_error_check(error, "read_block_connectivity")
     return conn
